third_party/blink/web_tests/svg/dynamic-updates/SVGLinearGradientElement-dom-y2-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y2' attribute of the SVGLinearGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS linearGradientElement.getAttribute('y2') is "1000%"
PASS linearGradientElement.getAttribute('y2') is "0%"
PASS successfullyParsed is true

TEST COMPLETE
